[FEAT]: Add seo best practices and readme.md file
