+++ README.md
@@ -1,36 +1,42 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
 
-## Getting Started
+## Correr Localmente
 
-First, run the development server:
+Clona el proyecto
 
 ```bash
-npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
+git clone https://github.com/RenzoDotJava/delosi-test
 ```
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+Dirigete al directorio
 
-You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
+```bash
+cd delosi-test
+```
 
-This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
+Instala las dependencias
 
-## Learn More
+```bash
+npm install
+```
 
-To learn more about Next.js, take a look at the following resources:
+Corre el proyecto
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+```bash
+npm run start
+```
 
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
+Ingresa una matriz NxN (Ej: [[1,2,3,4],[5,6,7,8],[9,10,11,12],[13,14,15,16]]) y obtén tu resultado
 
-## Deploy on Vercel
+![MatrixRotator Screenshots](https://pub-9b9cab9c477b4f089f886c993122867d.r2.dev/matrix_rotator_example.png)
 
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
 
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+
+## Corre los tests
+
+```bash
+npm run test
+```
+## Authors
+
+- [@renzogbo](https://www.github.com/RenzoDotJava)
+
