+++ src/__tests__/pages/MatrixRotatorPage.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, fireEvent, waitFor } from '@testing-library/react'
-import MatrixRotator from '@/app/page'
+import MainPage from '@/app/page'
 import {
   EMPTY_MATRIX,
   INVALID_NUMBER_MATRIX,
@@ -46,13 +46,13 @@ const getTextAreaByTestId = (testId: string) => screen.getByTestId(testId) as HT
 
 describe('when user visits the matrix rotator page', () => {
   it('should render the title', () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     expect(screen.getByText(MAIN_PAGE_TITLE)).toBeInTheDocument()
   })
 
   it('copy button should be disabled when there is no input matrix', () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     const copyButton = screen.getByText(/Copiar/)
 
@@ -61,7 +61,7 @@ describe('when user visits the matrix rotator page', () => {
 
   it('copy button should be enabled when there is an ouput matrix', () => {
     renderWithMatrixProvider({
-      children: <MatrixRotator />,
+      children: <MainPage />,
       mockState: {
         ...defaultMockState,
         matrix: '[[1,2],[3,4]]',
@@ -77,8 +77,8 @@ describe('when user visits the matrix rotator page', () => {
     expect(defaultMockState.copyToClipboard).toHaveBeenCalled()
   })
 
-  it('loading state should be displayed when user types a valid 4x4 matrix', async () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+  it('loading state should be displayed when user types a valid matrix', async () => {
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     const input = getTextAreaByTestId('input-matrix')
 
@@ -94,7 +94,7 @@ describe('when user visits the matrix rotator page', () => {
   })
 
   it('when user types a valid 4x4 matrix, it should render the rotated matrix', async () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     const input = getTextAreaByTestId('input-matrix')
     const output = getTextAreaByTestId('output-matrix')
@@ -119,7 +119,7 @@ describe('when user visits the matrix rotator page', () => {
 
 
   it('when user types a matrix with a size over MAX_SIZE_CANVAS, it should show MAX_SIZE_CANVAS_EXCEEDED state', async () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     const input = getTextAreaByTestId('input-matrix')
     const output = getTextAreaByTestId('output-matrix')
@@ -140,7 +140,7 @@ describe('when user visits the matrix rotator page', () => {
   })
 
   it('when user types a matrix with a size over MAX_SIZE_MATRIX, it should show OVER_MAX_SIZE_MATRIX state', async () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     const input = getTextAreaByTestId('input-matrix')
     const output = getTextAreaByTestId('output-matrix')
@@ -161,7 +161,7 @@ describe('when user visits the matrix rotator page', () => {
   })
 
   it('when user types an empty matrix, it should show EMPTY_MATRIX state', async () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     const input = getTextAreaByTestId('input-matrix')
     const output = getTextAreaByTestId('output-matrix')
@@ -180,7 +180,7 @@ describe('when user visits the matrix rotator page', () => {
   })
 
   it('when user types an invalid NxN matrix, it should show INVALID_NXN_MATRIX state', async () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     const input = getTextAreaByTestId('input-matrix')
     const output = getTextAreaByTestId('output-matrix')
@@ -199,7 +199,7 @@ describe('when user visits the matrix rotator page', () => {
   })
 
   it('when user types an invalid number matrix, it should show INVALID_NUMBER_MATRIX state', async () => {
-    renderWithMatrixProvider({ children: <MatrixRotator /> })
+    renderWithMatrixProvider({ children: <MainPage /> })
 
     const input = getTextAreaByTestId('input-matrix')
     const output = getTextAreaByTestId('output-matrix')
