--- src/components/pages/index.ts
@@ -1,1 +1,0 @@
-export { default as MainPage } from './main'
