--- src/components/pages/main/main.module.scss
@@ -1,87 +1,0 @@
-div.main-container {
-  display: flex;
-  min-height: 100vh;
-  flex-direction: column;
-  padding: 64px 200px;
-  gap: 36px;
-  transition: all 0.3s ease-in-out;
-
-  &>h1 {
-    text-align: center;
-  }
-
-  &>div:first-of-type{
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-  }
-
-  &>div:nth-of-type(2) {
-    display: flex;
-    gap: var(--subcontainer-gap);
-
-    &>div.field {
-      display: flex;
-      flex-direction: column;
-      width: var(--field-width);
-      max-height: 240px;
-      gap: 12px;
-
-      & > div {
-        display: flex;
-        justify-content: space-between;
-
-        & > button {
-          background-color: black;
-          color: white;
-          border: none;
-          padding: 0 16px;
-          border-radius: 4px;
-          cursor: pointer;
-          transition: all 0.1s ease-in-out;
-          font-family: var(--font-size-roboto);
-
-          &:disabled {
-            background-color: gray;
-            opacity: 0.8;
-            cursor: auto;
-          }
-
-          &:hover {
-            opacity: 0.8;
-          }
-        }
-      }
-
-      label {
-        font-size: var(--font-size-lg);
-        font-weight: bold;
-      }
-    }
-  }
-}
-
-@media only screen and (max-width: 1280px) {
-  div.main-container {
-    padding: 64px 50px;
-  }
-}
-
-@media only screen and (max-width: 1024px) {
-  div.main-container {
-    padding: 64px 25px;
-  }
-}
-
-@media only screen and (max-width: 768px) {
-  div.main-container {
-
-    &>div:nth-of-type(2){
-      flex-direction: column;
-
-      &>div.field {
-        width: 100%;
-      }
-    }
-  }
-}
