--- src/components/templates/index.ts
@@ -1,1 +1,0 @@
-export { default as Matrices } from './matrices'
