+++ src/constants/index.ts
@@ -1,4 +1,5 @@
 export const MAIN_PAGE_TITLE = 'Rotador de matrices NxN'
+export const MAIN_PAGE_DESCRIPTION = 'Este programa te permitirá rotar una matriz NxN 90 grados a la izquierda. Podrás copiar la salida y ver en pantalla tanto la matriz de entrada como rotada.'
 export const MAX_SIZE_CANVAS = 10
 export const MAX_SIZE_MATRIX = 1000
 
@@ -11,3 +12,5 @@ export const INVALID_NXN_MATRIX = 'Matriz inválida: El input debe ser una matri
 export const EMPTY_MATRIX = 'Matriz vacía: El input debe tener elementos'
 export const INVALID_NUMBER_MATRIX = 'Matriz inválida: Todos los elementos deben ser números'
 export const OVER_MAX_SIZE_MATRIX = 'Nuestros servidores no pueden procesar una matriz tan grande'
+
+export const BASE_URL = 'https://matrixrotator.com' //Should be on the env file
